change GetSpeed return type from int to float

+++ SDL/Enemy.cpp
@@ -106,7 +106,7 @@ int Enemy::GetHP() const {
 int Enemy::GetMaxHP() const { 
     return maxHealth; }
 
-int Enemy::GetSpeed() const {
+float Enemy::GetSpeed() const {
     return speed;
 }
